+++ examples/next/app/[[...slug]]/page.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { notFound } from "next/navigation";
 
 import { db } from "~/server/db";
-import { StaticContentRenderer } from "./static-content-renderer";
+import { BuilderContentRenderer } from "./builder-content-renderer";
 
 export async function generateStaticParams() {
   return [{ slug: ["/"] }];
@@ -19,5 +19,5 @@ export default async function Home({ params }: { params: { slug: string[] } }) {
     return notFound();
   }
 
-  return <StaticContentRenderer content={page.content} />;
+  return <BuilderContentRenderer content={page.content} />;
 }