+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @simple-builder/react
 
+## 1.0.5
+
+### Patch Changes
+
+- allow usage of custom default styles in builder registery
+- Updated dependencies
+  - @simple-builder/server@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes